go back to Stan.h rather than StanVar.h to make it build
There's some include-ordering-problem here with the initialization of Eigen vs Stan's injected type-traits stuff.  I wasn't able to figure out what was missing in the previous version, so for now reverting to #including the whole Stan stack.
This isn't great because it significantly slows down building, but at least seems to work for the moment.