Update Hist.h to account for case where fSqrtErrs = true.
Sorry this has taken me so long to get back to! Is this more like what you wanted? I was not sure whether I should set the existing array fSumSq to the sqrt errors if fSqrtErrs=true, or if I should create a new array and return that. The function is quite long now as well, so would you prefer it if I put the implementation into the .cxx file instead?